requirements.txt
streamlit
ultralytics
torch
pillow
pandas
opencv-python
